Fix wrong variable type in MP_MQ2 and MP_MQ3

+++ lib/arduino/MP_MQ3/MP_MQ3.h
@@ -16,8 +16,7 @@ public:
 
 private:
 	uint8_t pin;
-	int percent = 0;
-	
+	double percent = 0;
 };
 
 #endif
